requirements.txt
streamlit==1.31.0
google-generativeai==0.3.2
sentence-transformers==2.3.1
PyPDF2==3.0.1
torch==2.1.2